Modify EO class to use new shared model paradigm for EnumOutputs.
The EO class now directly populates the shared model then updates attached
widgets with that model. This allows for more complicated enum controls and immediately
allows for the addition of tooltips to options in the pull-down menu.

This also implements the default_value control of the EO, so if settings does not
have a setting, the specified default value is used instead.

This also allows for specifying the parent object of the EO widgets.

Finally, property AO,EO widgets now default to using their short name as the label.